% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dashboard.R
\name{add_dashboard}
\alias{add_dashboard}
\title{Add a new dashboard}
\usage{
add_dashboard(name = NULL, baseurl = NULL)
}
\arguments{
\item{name}{character string, name of the dashboard to add}

\item{baseurl}{character string, the server to add the dashboard}
}
\value{
S3 object with content, path and the response of the dashboard
}
\description{
Add a new dashboard
}
