% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_content}
\alias{check_content}
\title{check the api returned json}
\usage{
check_content(res)
}
\arguments{
\item{res}{A character, the http response}
}
\value{
An error if content is not json
}
\description{
check the api returned json
}
