% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixthis-package.R
\docType{package}
\name{fixthis-package}
\alias{fixthis}
\alias{fixthis-package}
\title{fixthis: Fix Common Issues With PSI-MIS}
\description{
Use this to fix commonly known issues at PSI
    freshdesk.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://this}
  \item Report bugs at \url{http://that}
}

}
\author{
\strong{Maintainer}: Isaiah Nyabuto \email{isaiahnyabuto@gmail.com}

}
\keyword{internal}
