% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{get_resources}
\alias{get_resources}
\title{Retrieve a resource}
\usage{
get_resources(url = "", type = NULL, id = NULL, ...)
}
\arguments{
\item{url}{a character string (or more extended vector, e.g., for the
"libcurl." method) naming the base URL of the server.}

\item{type}{a character string with the name of a resource to pull. e.g data
elements, maps, charts. By default it will retrieves all the resource end
points available at PSI - MIS.}

\item{id}{a character string, the id of a specific resource to pull.}

\item{...}{allow additional arguments to be passed. Example a filter, or a
query.}
}
\value{
an S3 type object with the content of the resource, path and the
response
}
\description{
This is function pulls a resource at PSI - MIS web API. It returns a list the
content, path, and the response.
}
