% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unshared_vis.R
\name{get_unshared_vis}
\alias{get_unshared_vis}
\title{Get a list of unshared visualization}
\usage{
get_unshared_vis(
  owner_id = NULL,
  type = NULL,
  baseurl = NULL,
  paging = TRUE,
  startsWith = NULL,
  createdLaterThan = NULL,
  ...
)
}
\arguments{
\item{owner_id}{character string, user id to dig.}

\item{type}{character string, type of visualization. (maps, charts, tables).}

\item{baseurl}{character string, the server url.}

\item{...}{additional parameters passed to \code{get_resources}.}
}
\value{
list with unshared visualization.
}
\description{
Retrieve unshared visualization (maps, charts and tables) by date.
}
\author{
Isaiah Nyabuto
}
