% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vis_owner.R
\name{get_vis_owner}
\alias{get_vis_owner}
\title{Get the owner of a visualization}
\usage{
get_vis_owner(visual, type = NULL, baseurl)
}
\arguments{
\item{visual}{a data frame with visualization ids.}

\item{type}{a character vector indicating the type of visualization. Example;
maps, charts or reportTables.}

\item{baseurl}{a character vector specifying a URL of the server.}

\item{...}{Additional parameters passed to get_resources API.}
}
\value{
A list of user names and ids.
}
\description{
This function gets the owner of a visualization. it's used in the
\code{dash_builders} to extract a list of dashboard collaborators
}
