% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_multiple_ids}
\alias{has_multiple_ids}
\title{checks if the visaulization has muiltiple ids}
\usage{
has_multiple_ids(visual)
}
\arguments{
\item{visaul}{A data frame with visualization ids.}
}
\value{
logical output
}
\description{
checks if the visaulization has muiltiple ids
}
