% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_not_supported.R
\name{items_not_supported}
\alias{items_not_supported}
\title{Recover unsupported items in a dashboard}
\usage{
items_not_supported(dashboard, dashboard_id, ...)
}
\arguments{
\item{dashboard}{a character string, name the dashboard with unsupported
items.}

\item{dashboard_id}{character, id of the dashboard with unsupported items.
This is an optional field; you do not need to specify this if your have already specified the correct name of the dashboard.}

\item{...}{Additional parameters passed to items_not_supported}
}
\description{
This function is used to retrieve unsupported items on a dashboard. These are
items that commonly disappear when a user with limited access on a dashboard
saves changes on the dashboard.
}
