% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vis_ids}
\alias{vis_ids}
\title{Pop visualisation ids from a visual (df)}
\usage{
vis_ids(visual)
}
\arguments{
\item{visaul}{A data frame with visualization ids.}
}
\value{
A character vector with ids of the visualization.
}
\description{
Pop visualisation ids from a visual (df)
}
