% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{add_dashboard}
\alias{add_dashboard}
\alias{add_dashboard_items}
\title{Add a new dashboard}
\usage{
add_dashboard(name = NULL, baseurl = NULL)

add_dashboard_items(dashboard_id = "", parsed = list(), baseurl = NULL)
}
\arguments{
\item{name}{character string, name of the dashboard to add}

\item{baseurl}{character string, the server to add the dashboard}

\item{parsed}{A list, content of the unshared visualization.}
}
\value{
S3 object with content, path and the response of the dashboard
}
\description{
query <- paste0("type=", type(names(parsed$content\link{2})), "&id=", parsed$content[\link{2}]$id)
}
\section{item_type <- paste0("items", type(names(parsed$content\link{2})), sep = "/")}{
}
\section{item_id <-}{
}

